docs(general): clearer `sprite_flush_multi` example
The original example uses the undeclared variable `spr_a`, relying on the obscure behaviour of unset variables implicitly convertible to arrays, and then assigns indices one-by-one.

I've changed this to defining an instance variable with a more relevant name that's passed to the function. This could be made inline, but the first line could be put in the create event and re-used.